# backend/api-spec-complete.yaml
openapi: 3.0.3
info:
  title: RetailGenie API
  description: |
    **RetailGenie** - Comprehensive retail management API with AI-powered recommendations,
    inventory management, user authentication, and real-time analytics.

    ## 🚀 Features
    - 🛍️ **Product Management**: Full CRUD operations for product catalog
    - 🔐 **Authentication**: Secure user registration and login system
    - 🤖 **AI Assistant**: Intelligent product recommendations and chat
    - 📊 **Analytics**: Real-time business insights and dashboards
    - 💰 **Pricing**: Dynamic pricing optimization and competitor analysis
    - 📦 **Inventory**: Stock tracking and management
    - 💬 **Feedback**: Customer reviews and rating system
    - 🔌 **WebSocket**: Real-time notifications and updates
    - 🔄 **Background Tasks**: Asynchronous processing with Celery

    ## 🔐 Authentication
    Most endpoints require authentication via Bearer token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## 📈 API Versioning
    This API supports multiple versions:
    - **v1**: Stable production API (current)
    - **v2**: Enhanced API with advanced features

    Use version-specific endpoints: `/api/v1/...` or `/api/v2/...`

    ## ⚡ Rate Limiting
    API requests are limited to:
    - **1000 requests/hour** for authenticated users
    - **100 requests/hour** for unauthenticated users

    ## 📄 Pagination
    List endpoints support pagination with standard parameters:
    - `page`: Page number (default: 1)
...

...

...

...
servers:
...
paths:
...
  /api/v1/products:
    get:
      summary: Get all products (v1)
      description: Retrieve a paginated list of all products with optional filtering
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/InStockFilter'
        - $ref: '#/components/parameters/MinPriceFilter'
        - $ref: '#/components/parameters/MaxPriceFilter'
      responses:
        '200':
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new product (v1)
      description: Create a new product in the catalog
      tags:
        - Products
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/products/{product_id}:
    get:
      summary: Get product by ID (v1)
      description: Retrieve a specific product by its ID
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update product (v1)
      description: Update an existing product
      tags:
        - Products
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete product (v1)
      description: Delete a product from the catalog
      tags:
        - Products
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '204':
          description: Product deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Enhanced V2 Product Endpoints
  /api/v2/products:
    get:
      summary: Get all products (v2)
      description: Enhanced product listing with advanced filtering, sorting, and analytics
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/InStockFilter'
        - $ref: '#/components/parameters/MinPriceFilter'
        - $ref: '#/components/parameters/MaxPriceFilter'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - $ref: '#/components/parameters/IncludeAnalyticsParam'
      responses:
        '200':
          description: Enhanced product list with metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new product (v2)
      description: Create a new product with enhanced features and validation
      tags:
        - Products
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnhancedProductCreateRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedProduct'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v2/products/{product_id}:
    get:
      summary: Get product by ID (v2)
      description: Retrieve a specific product with enhanced data and recommendations
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/IncludeRecommendationsParam'
        - $ref: '#/components/parameters/IncludeAnalyticsParam'
      responses:
        '200':
          description: Enhanced product data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedProduct'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Search Endpoints
...
  /api/auth/register:
    post:
      summary: Register new user
      description: Create a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user and return JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # AI Assistant Endpoints
...
  /api/feedback/{product_id}:
    get:
      summary: Get product feedback
      description: Retrieve all feedback and reviews for a specific product
      tags:
        - Feedback
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Product feedback retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/feedback:
    post:
      summary: Submit product feedback
      description: Submit a new review or feedback for a product
      tags:
        - Feedback
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '201':
          description: Feedback submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Admin Endpoints
  /api/admin/init-db:
    post:
      summary: Initialize database
      description: Initialize database with sample data (development only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Database initialized successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Database initialized successfully"
                  products_created:
                    type: integer
                    example: 3
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # WebSocket Info Endpoint
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    ProductIdParam:
      name: product_id
      in: path
      required: true
      description: Unique identifier for the product
      schema:
        type: string
        example: "product-123"
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    CategoryFilter:
      name: category
      in: query
      description: Filter by product category
      schema:
        type: string
        example: "Electronics"
    SearchQuery:
      name: search
      in: query
      description: Search products by name or description
      schema:
        type: string
        example: "wireless"
    InStockFilter:
      name: in_stock
      in: query
      description: Filter by stock availability
      schema:
        type: boolean
        example: true
    MinPriceFilter:
      name: min_price
      in: query
      description: Minimum price filter
      schema:
        type: number
        format: float
        minimum: 0
        example: 10.00
    MaxPriceFilter:
      name: max_price
      in: query
      description: Maximum price filter
      schema:
        type: number
        format: float
        minimum: 0
        example: 500.00
    SortByParam:
      name: sort_by
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [name, price, created_at, updated_at, rating]
        default: created_at
        example: price
    SortOrderParam:
      name: sort_order
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
        example: asc
    IncludeRecommendationsParam:
      name: include_recommendations
      in: query
      description: Include product recommendations
      schema:
        type: boolean
        default: false
        example: true
    IncludeAnalyticsParam:
      name: include_analytics
      in: query
      description: Include analytics data
      schema:
        type: boolean
        default: false
        example: true
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: string
          description: Unique product identifier
          example: "product-123"
        name:
          type: string
          description: Product name
          example: "Wireless Bluetooth Headphones"
        price:
          type: number
          format: float
          description: Product price in USD
          minimum: 0
          example: 79.99
        category:
          type: string
          description: Product category
          example: "Electronics"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones with noise cancellation"
        in_stock:
          type: boolean
          description: Whether the product is in stock
          example: true
        stock_quantity:
          type: integer
          description: Current stock quantity
          minimum: 0
          example: 50
        created_at:
          type: string
          format: date-time
          description: Product creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Product last update timestamp
        version:
          type: string
          description: API version that created this response
          example: "1.0.0"
...
    ProductCreateRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          description: Product name
          example: "Wireless Bluetooth Headphones"
        price:
          type: number
          format: float
          description: Product price in USD
          minimum: 0
          example: 79.99
        category:
          type: string
          description: Product category
          example: "Electronics"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones"
        in_stock:
          type: boolean
          description: Whether the product is in stock
          default: true
        stock_quantity:
          type: integer
          description: Initial stock quantity
          minimum: 0
          default: 0
...
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Product name
        price:
          type: number
          format: float
          description: Product price in USD
          minimum: 0
        category:
          type: string
          description: Product category
        description:
          type: string
          description: Product description
        in_stock:
          type: boolean
          description: Whether the product is in stock
        stock_quantity:
          type: integer
          description: Stock quantity
          minimum: 0
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        count:
          type: integer
          description: Total number of products returned
          example: 25
        total:
          type: integer
          description: Total number of products matching criteria
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        pages:
          type: integer
          description: Total number of pages
          example: 8
        version:
          type: string
          example: "1.0.0"
    EnhancedProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/EnhancedProduct'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        filters:
          $ref: '#/components/schemas/AppliedFilters'
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'

    # Search Schemas
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input data"
            status_code: 400
            timestamp: "2023-01-15T10:30:00Z"
            path: "/api/products"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            status_code: 401
            timestamp: "2023-01-15T10:30:00Z"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Insufficient permissions"
            status_code: 403
            timestamp: "2023-01-15T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            status_code: 404
            timestamp: "2023-01-15T10:30:00Z"
            path: "/api/products/invalid-id"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            status_code: 500
            timestamp: "2023-01-15T10:30:00Z"

# Example webhook configurations for real-time updates
...